test: Remove deprecated CSP integration tests
Delete the CSP integration tests along with every CSP reference in the
test infrastructure: the two SecurityPoliciesMostRestrictive tests, the
nine CSPDisablesForwardingTests logging classes, the four CSP fixtures,
and the securityPoliciesToken plumbing in RemoteApplicationFixture. No
test now uses the CSP license key.

Rename the CSP test folder and namespaces to HSM to reflect what they
actually cover - 11 of the 13 files in tests/.../IntegrationTests/CSP
were High Security Mode tests.

This replaces the earlier SkipWhen approach, which failed at runtime.
xunit v3 resolves the named property with GetProperty(Public | Static)
against the concrete test class and without FlattenHierarchy, so the
static IsCspFixture declared on the shared generic base was never
found and all 17 Logging.HsmAndCsp tests errored, HSM variants
included.

The agent's own security-policies support is unchanged; the
Supportability/FeatureEnabled/CSP metric and its unit test remain.
