+++ tests/Agent/IntegrationTests/IntegrationTests/HSM/HighSecurityModeNoTransactionAgentApiTests.cs
@@ -9,7 +9,7 @@ using NewRelic.Agent.Tests.TestSerializationHelpers.Models;
 using NewRelic.Testing.Assertions;
 using Xunit;
 
-namespace NewRelic.Agent.IntegrationTests.CSP;
+namespace NewRelic.Agent.IntegrationTests.HSM;
 
 public class HighSecurityModeNoTransactionAgentApiTests : NewRelicIntegrationTest<RemoteServiceFixtures.HSMAgentApiExecutor>
 {
