+++ tests/Agent/IntegrationTests/UnboundedServices/example-secrets.json
@@ -15,9 +15,6 @@
       "HSM": {
         "LicenseKey": "YOUR_NEW_RELIC_HSM_KEY_HERE"
       },
-      "CSP": {
-        "LicenseKey": "YOUR_NEW_RELIC_CSP_KEY_HERE"
-      },
       "PostgresTests": {
         "CustomSettings": {
           "ConnectionString": "Server=127.0.0.1;Port=5432;User Id=postgres;Password=PostgresPassword;Database=postgres;"
